packages/metadata-client/metadata_client-3.0.8.tar.gz/metadata_client-3.0.8/metadata_client/apis/data_group_repository_api.py
"""DataGroupRepositoryApi module class"""

import json

from ..common.base import Base


class DataGroupRepositoryApi(Base):
    def create_data_group_repository_api(self, data_group_repository):
        api_url = self.__get_api_url()
        return self.api_post(api_url, data=json.dumps(data_group_repository))

    def delete_data_group_repository_api(self, data_group_repository_id):
        api_url = self.__get_api_url(data_group_repository_id)
        return self.api_delete(api_url)

    def update_data_group_repository_api(self, data_group_repository_id,
                                         data_group_repository):
        api_url = self.__get_api_url(data_group_repository_id)
        return self.api_put(api_url, data=json.dumps(data_group_repository))

    def get_data_group_repository_by_id_api(self, data_group_repository_id):
        api_url = self.__get_api_url(data_group_repository_id)
        return self.api_get(api_url, params={})

    def get_all_by_data_group_id_api(self, data_group_id):
        api_url = self.__get_api_url()
        return self.api_get(api_url, params={'data_group_id': data_group_id})

    def get_all_by_data_group_id_and_repository_id_api(self, data_group_id,
                                                       repository_id):
        api_url = self.__get_api_url()
        return self.api_get(api_url, params={'data_group_id': data_group_id,
                                             'repository_id': repository_id})

    #
    # Private Methods
    #
    def __get_api_url(self, api_specifics=''):
        model_name = 'data_groups_repositories/'
        return self.get_api_url(model_name, api_specifics)
